requirements.txt
arcgis
pandas
python-dateutil
python-dotenv
requests
